socket_server/config.py
import os
from dotenv import load_dotenv

load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL")
SERVER_HOST = os.getenv("SERVER_HOST")
SERVER_PORT = int(os.getenv("SERVER_PORT", 8000))
FIREBASE_PROJECT_ID = os.getenv("FIREBASE_PROJECT_ID")
