socket_server/requirements.txt
protobuf
psycopg2-binary
python-dotenv
websockets
google-auth
requests
